requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openai>=1.0.0
httpx>=0.25.0
tenacity>=8.2.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
pypdf>=3.0.0
python-docx>=0.8.11
